+++ CHANGELOG.md
@@ -14,3 +14,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Create endpoint to add new recipe
 - Accept both tags and labels parameters in DTO from firebase
 - Retrieve/accept time and url params
+- Create endpoint to retrieve recipe by id